log final error message as a single string
For interactive workflows the UI shows the final log from the connector as the
error, and specifically for Discover it will show the "message" part of the log
in the UI prominently.

Prior to this change, that would just be the "network-tunnel failed." part, even
though there was more information in the structured log.

This modifies the log output to include the actual error message in the message.